api/requirements.txt
tensorflow
pillow
tensorflow-serving-api
matplotlib
numpy
flask
requests